test(igxFocusTrap): split igxFocusTrap tests#8961
